backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4